+++ src/fullctl/django/rest/mixins.py
@@ -12,6 +12,32 @@ class CachedObjectMixin:
         return self._obj
 
 
+class SlugObjectMixin:
+
+    """
+    Assures that a views get_object() can do a lookup
+    via either the slug or the id
+    """
+
+    slug_field = "slug"
+
+    def get_object(self):
+        """
+        pk could be either slug or id
+        """
+
+        lookup_value = self.kwargs.get(self.lookup_field)
+        org_tag = self.kwargs.get("org_tag")
+
+        if not lookup_value:
+            return None
+
+        if lookup_value.isdigit():
+            return self.queryset.get(instance__org__slug=org_tag, id=lookup_value)
+        else:
+            return self.queryset.get(instance__org__slug=org_tag, slug=lookup_value)
+
+
 class OrgQuerysetMixin:
     """
     For objects with URLs that require an "org_tag", this filters